Makes Sidebar Items Dissapear when Updates
When the version of the app is the same as the version of cobalt
there is no need to show a list of all components that are with
the empty state of checked.

There fore, this commit makes it that the sidebar turns into an
empty state which shows the user the app is updated

The design could be better.